requirements.txt
asyncio==3.4.3
aiohttp>=3.9.0
openai>=1.0.0
python-dotenv>=1.0.0 